+++ tests/functional/test_plugins.py
@@ -158,7 +158,7 @@ def test_aws_no_imds(
             config=CONFIGS / "imds.c4m",
             env={
                 "VENDOR": str(vendor),
-                "METADATA_IP": "127.0.0.1",
+                "METADATA_IP": "imds",
                 "INSTANCE": str(instance),
             },
         )
@@ -167,6 +167,7 @@ def test_aws_no_imds(
                 "_OP_CLOUD_SYS_VENDOR": "Amazon EC2",
                 "_OP_CLOUD_PROVIDER": "aws",
                 "_OP_CLOUD_PROVIDER_SERVICE_TYPE": "aws_ec2",
+                "_OP_CLOUD_METADATA_FAILURE_REASON": re.compile(r"disabled"),
                 "_AWS_INSTANCE_ID": "i-abc123xyz789",
             }
         )